Facturas telefono y correccion BBDD

+++ es.upm.dit.isst.iFactura2016/src/es/upm/dit/isst/iFactura2016/model/FacturaLuz.java
@@ -22,7 +22,7 @@ public class FacturaLuz implements Serializable {
 	private Long id;
 
 	/** The cliente. */
-	private Integer cliente;
+	private Long cliente;
 
 	/** The empresa. */
 	private String empresa;
@@ -75,7 +75,7 @@ public class FacturaLuz implements Serializable {
 	 *
 	 * @return the cliente
 	 */
-	public Integer getCliente() {
+	public Long getCliente() {
 		return cliente;
 	}
 
@@ -85,7 +85,7 @@ public class FacturaLuz implements Serializable {
 	 * @param cliente
 	 *            the new cliente
 	 */
-	public void setCliente(Integer cliente) {
+	public void setCliente(Long cliente) {
 		this.cliente = cliente;
 	}
